super janky end to end oidc, so much cleanup needed, but it goes e2e
extracting the aws stuf as it is really mucking up the repo. I had thought having it in the core
would be helpful, but the sprawl of terraform, configs, etc speaks to making a template repo
instead of putting it in this repo, so that folks can clone the template and customize
instead of having to manage stuff *in* this repo, which is justy a bad experience.